+++ src/config.js
@@ -19,7 +19,7 @@ const config = {
         default: false,
         order: 2
     },
-    panelPosition: {
+    'panelPosition': {
         title: 'Position of the panel',
         description: 'You can set the place of the viewer, to the most right position or to the most left position.',
         type: 'string',
@@ -27,23 +27,29 @@ const config = {
         enum: ['Left', 'Right'],
         order: 3
     },
+    'alwaysOpenInNewWindow': {
+        description: 'If set to true, always open projects in a new window (default Atom\'s behavior), instead of opening in the same window.',
+        type: 'boolean',
+        default: false,
+        order: 4
+    },
     'hideHeader': {
         description: 'Hide header (for more space).',
         type: 'boolean',
         default: false,
-        order: 4
+        order: 5
     },
     'githubToken': {
         description: 'Your personal and private GitHub token. This is useful if you want to save/backup your projects to a remote place (as a gist). *note*: keep in mind that this token should have only permissions to `rw` gists as well as that any package can access this token string.',
         type: 'string',
         default: '',
-        order: 5
+        order: 6
     },
     'convertOldData': {
         description: 'If you came from a version previous to <code>0.3.0</code>, you most probably have the old data in the atom folder. By default it will always check on startup for this data and if the new does not exist, it will convert to the new data schema.',
         type: 'boolean',
         default: true,
-        order: 6
+        order: 7
     }
 };